review polish: hoist tz_utils out of ingest/, fix docstring, log change
Three follow-ups from review of the timezone-centralization PR:

1. tz_utils.py moves from src/rebalance/ingest/ to src/rebalance/. It is
   imported from scripts/dashboard.py and from inside ingest/ — i.e. it
   is a cross-cutting utility, not an ingestor. New home is the sibling
   of paths.py, which is the existing cross-cutting helper.

2. The calendar_config.py top-of-file schema example previously used a
   JSON line with a "//" comment ("timezone": "America/Los_Angeles",
   // omit ... to use device local timezone"), which is illegal JSON
   and contradicted the stated default. Rewritten to show timezone: ""
   (matching DEFAULT_CONFIG) with the explanation in surrounding prose.

3. CHANGELOG gains a [0.28.2] entry noting the user-visible behavior
   change: timestamps now default to OS-detected local time instead of
   hardcoded America/Los_Angeles / America/New_York fallbacks.
