+++ scripts/dashboard.py
@@ -51,7 +51,7 @@ from rebalance.ingest.config import (  # noqa: E402
     get_pulse_config,
 )
 from rebalance.ingest.db import db_connection  # noqa: E402
-from rebalance.ingest.tz_utils import local_tz  # noqa: E402
+from rebalance.tz_utils import local_tz  # noqa: E402
 from rebalance.ingest.index_ops import (  # noqa: E402
     get_index_status,
     get_watched_repos,
